djGit/cassini/requirements.txt
astroid==2.2.5
colorama==0.4.1
decorator==4.1.2
defusedxml==0.6.0
Django==1.11.4
django-auditlog==0.4.5
django-composite-foreignkey==1.0.1
django-filter==1.0.4
django-jenkins==0.110.0
django-jsonfield==1.2.0
django-linter==0.7
djangorestframework==3.6.4
djangorestframework-xml==1.4.0
dnspython==1.15.0
lazy-object-proxy==1.3.1
pylint==1.5.1
python-dateutil==2.6.0
pytz==2017.2
PyYAML==5.1
pyzmq==18.0.1
six==1.12.0
typed-ast==1.3.5
wrapt==1.11.1
mysqlclient == 1.3.12
python-decouple==3.1
